fix: gemini core review
