+++ a2a-client/src/rest.rs
@@ -505,6 +505,7 @@ impl TransportFactory for RestTransportFactory {
 #[cfg(test)]
 mod tests {
     use super::*;
+    use crate::test_utils::rcgen_self_signed_ca_pem;
     use serde_json::json;
 
     #[test]
@@ -651,16 +652,5 @@ mod tests {
         let iface = AgentInterface::new("https://localhost:3443/rest", "HTTP+JSON");
         let transport = f.create(&card, &iface).await.unwrap();
         transport.destroy().await.unwrap();
-    }
-
-    fn rcgen_self_signed_ca_pem() -> Vec<u8> {
-        let mut params = rcgen::CertificateParams::new(Vec::<String>::new()).unwrap();
-        params.is_ca = rcgen::IsCa::Ca(rcgen::BasicConstraints::Unconstrained);
-        params
-            .distinguished_name
-            .push(rcgen::DnType::CommonName, "Test CA");
-        let key = rcgen::KeyPair::generate().unwrap();
-        let cert = params.self_signed(&key).unwrap();
-        cert.pem().into_bytes()
     }
 }
